@@ -1,5 +1,5 @@
-import 'package:airlux/screens/devices_screen.dart';
 import 'package:flutter/material.dart';
+import 'screens/login_screen.dart';
 
 void main() {
   runApp(const MyApp());
@@ -15,52 +15,75 @@ class MyApp extends StatelessWidget {
       title: 'Flutter Demo',
       theme: ThemeData(
         primarySwatch: Colors.deepPurple,
-      ),
-      home: const MyHomePage(title: 'Accueil'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'Bonjour!',
-              style: Theme.of(context).textTheme.headline4,
-            ),
-          ],
+        textTheme: const TextTheme(
+          titleLarge: TextStyle(fontSize: 36.0, fontWeight: FontWeight.bold, fontStyle: FontStyle.italic),
+          titleMedium: TextStyle(fontSize: 20.0, fontWeight: FontWeight.bold),
+          bodyMedium: TextStyle(fontSize: 16.0),
+          bodySmall: TextStyle(fontSize: 14.0),
         ),
+        // appBarTheme: AppBarTheme(
+        //   bo
+        // ),
+        inputDecorationTheme: const InputDecorationTheme(
+          enabledBorder: OutlineInputBorder(
+            borderSide: BorderSide(color: Colors.black),
+            borderRadius: BorderRadius.all(Radius.circular(100)),
+          ),
+          focusedBorder: OutlineInputBorder(
+            borderSide: BorderSide(color: Colors.deepPurple),
+          ),
+          filled: true,
+          hintStyle: TextStyle(fontSize: 16.0, fontWeight: FontWeight.normal),
+        ),
+        elevatedButtonTheme: ElevatedButtonThemeData(
+          style: ElevatedButton.styleFrom(
+            foregroundColor: Colors.white,
+            textStyle: const TextStyle(fontSize: 20),
+            backgroundColor: Colors.deepPurple,
+            shape: const StadiumBorder(),
+            ),
+        ),
+        textButtonTheme: TextButtonThemeData(
+          style: TextButton.styleFrom(
+            foregroundColor: Colors.deepPurple,
+            textStyle: Theme.of(context).textTheme.bodySmall,
+          ))
       ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-                builder: (context) => const DevicesScreen(
-                      title: 'Appairage',
-                    )),
-          );
-        },
-        tooltip: 'Add new devices',
-        child: const Icon(Icons.add),
-      ),
+      // home: const MyHomePage(title: 'Flutter Demo Home Page'),
+      home: LoginScreen(),
     );
   }
 }
+
+// class MyHomePage extends StatefulWidget {
+//   // const MyHomePage({super.key, required this.title});
+//   const MyHomePage({super.key});
+
+//   // final String title;
+
+//   @override
+//   State<MyHomePage> createState() => _MyHomePageState();
+// }
+
+// class _MyHomePageState extends State<MyHomePage> {
+
+//   @override
+//   Widget build(BuildContext context) {
+
+//     return Scaffold(
+//       // appBar: AppBar(
+//       //   title: Text(widget.title),
+//       // ),
+//       body: Center(
+//         child: Column(
+//           mainAxisAlignment: MainAxisAlignment.center,
+//           children: const <Widget>[
+//             Text(
+//               'Bonjour !',
+//             ),
+//           ],
+//         ),
+//       ),
+//     );
+//   }
+// }
